# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Poufmaker API",
    "version": "1.0.0",
    "description": "API documentation for Poufmaker application",
    "contact": {
      "name": "API Support",
      "email": "support@poufmaker.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string"
          },
          "Email": {
            "type": "string",
            "format": "email"
          },
          "FullName": {
            "type": "string"
          },
          "PhoneNumber": {
            "type": "string"
          },
          "Role": {
            "type": "string",
            "enum": [
              "client",
              "admin",
              "upholsterer"
            ]
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "Title": {
            "type": "string"
          },
          "Description": {
            "type": "string"
          },
          "Price": {
            "type": "number",
            "format": "float"
          },
          "ImageUrl": {
            "type": "string",
            "format": "uri"
          },
          "Status": {
            "type": "string",
            "enum": [
              "ai-generated",
              "pending",
              "approved",
              "rejected"
            ],
            "default": "ai-generated"
          },
          "CreatorId": {
            "type": "string",
            "format": "uuid"
          },
          "ManufacturerId": {
            "type": "string",
            "format": "uuid"
          },
          "CreatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "UpdatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "users_products_CreatorIdTousers": {
            "$ref": "#/components/schemas/User"
          },
          "users_products_ManufacturerIdTousers": {
            "$ref": "#/components/schemas/User"
          },
          "bids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bid"
            }
          }
        }
      },
      "Bid": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid"
          },
          "ProductId": {
            "type": "string",
            "format": "uuid"
          },
          "UpholstererId": {
            "type": "string",
            "format": "uuid"
          },
          "Amount": {
            "type": "number",
            "format": "float"
          },
          "Status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "rejected"
            ],
            "default": "pending"
          },
          "Notes": {
            "type": "string"
          },
          "CreatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "UpdatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List products",
        "description": "Get a list of products with optional filtering and pagination",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Items per page"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "ai-generated",
                "pending",
                "approved",
                "rejected"
              ]
            },
            "description": "Filter by status"
          },
          {
            "in": "query",
            "name": "creatorId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by creator ID"
          },
          {
            "in": "query",
            "name": "manufacturerId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter by manufacturer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        },
                        "currentPage": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create product",
        "description": "Create a new product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "Title"
                ],
                "properties": {
                  "Title": {
                    "type": "string",
                    "example": "Modern Pouf"
                  },
                  "Description": {
                    "type": "string",
                    "example": "A comfortable modern pouf with premium fabric"
                  },
                  "Price": {
                    "type": "number",
                    "format": "float",
                    "example": 99.99
                  },
                  "ImageUrl": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://example.com/pouf.jpg"
                  },
                  "Status": {
                    "type": "string",
                    "enum": [
                      "ai-generated",
                      "pending",
                      "approved",
                      "rejected"
                    ],
                    "default": "ai-generated"
                  },
                  "ManufacturerId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}